+++ back-end/readme.txt
@@ -1,0 +1,3 @@
+The back-end of your project will live in this directory.
+
+You will most likely initiate the back-end Node.js/Express.js server by running the "npm init" command from within this directory.
