--- front-end/src/components/student/CreateRequest.css
@@ -1,155 +1,0 @@
-@import url("../../assets/styles.css");
-
-
-
-.create-request-component-student {
-    height: 600px;
-    width: 500px;
-    position: absolute;
-    border: 1px solid #000; /* Optional border for visualization */
-    background-color: #fff;
-    top: 50%; /* Move the container down by 50% of the parent's height */
-    left: 50%; /* Center the container horizontally by 50% of the parent's width */
-    transform: translate(-50%, -50%); /* Center the container precisely */
-}
-.create-request-component-student .top-rectangle {
-    background-color: #57068C; /* Set the background color to #57068C */
-    width: 500px; /* Set the width to 637px */
-    height: 50px; /* Set the height to 50px */
-    position: absolute;
-    top: 0; /* Set the Y coordinate to 0 */
-    left: 0; /* Set the X coordinate to 0 */
-}
-
-.create-request-component-student .create-request-title {
-    color: #57068C; 
-    font-size: 45px; 
-    margin-bottom: 10px; /* Add some space below the title */
-    position: absolute; 
-    top: 50px; /* Set the Y coordinate to 63px */
-    left: 50px; /* Set the X coordinate to 63px */
-    font-family: var(--primary-font), var(--secondary-font);
-
-  }
-
-.create-request-component-student .issue-title-group {
-    position: absolute;
-    top: 160px; /* Set the Y coordinate to 162px */
-    left: 65px; /* Set the X coordinate to 75px */
-}
-
-/* Style for the label "Issue Title" */
-.create-request-component-student .issue-title-label {
-    color: #3D7D2E; /* Set the text color to #3D7D2E */
-    font-size: 24px; 
-    font-family: var(--primary-font), var(--secondary-font);
-}
-
-/* Style for the text input */
-.create-request-component-student .issue-title-input {
-    position: absolute;
-    width: 350px; /* Adjust the width of the text box as needed */
-    margin-top: 30px;
-    left: 1px; /* Set the X coordinate to 75px */
-    
-}
-
-
-.create-request-component-student .description-group {
-    position: absolute;
-    top: 230px; /* Set the Y coordinate to 192px */
-    left: 65px; /* Set the X coordinate to 75px */
-}
-
-/* Style for the label "Description" */
-.create-request-component-student .description-label {
-    color: #3D7D2E; /* Set the text color to #3D7D2E */
-    font-size: 24px; 
-    font-family: var(--primary-font), var(--secondary-font);
-    margin-bottom: 5px; /* Add some space below the title */
-}
-
-/* Style for the text area input */
-.create-request-component-student .description-input {
-    width: 350px; /* Adjust the width of the text area as needed */
-    height: 100px; /* Adjust the height of the text area as needed */
-    position: absolute;
-    left:1px;
-    margin-top: 10px;
-}
-
-
-
-
-.create-request-component-student .department-group {
-    position: absolute;
-    top: 390px; /* Adjust the Y coordinate as needed */
-    left: 65px; /* Set the X coordinate to 75px (or adjust as needed) */
-}
-
-/* Style for the label "Select Department" */
-.create-request-component-student .department-label {
-    color: #3D7D2E; /* Set the text color to #57068C or your desired color */
-    font-size: 18px; /* Adjust the font size as needed */
-    font-family: var(--primary-font), var(--secondary-font);
-}
-
-/* Style for the department selection dropdown */
-.create-request-component-student .department-select {
-    width: 200px; /* Adjust the width of the dropdown as needed */
-    position: absolute;
-    margin-top: 10px;
-    left: 1px;
-}
-
-
-.create-request-component-student .file-group{
-position:absolute;
-    top: 390px;
-    left: 300px;
-}
-
-.create-request-component-student .file-label{
-    color: #29561f; /* Set the text color to #57068C or your desired color */
-    font-weight: 500;
-    font-size: 15px; /* Adjust the font size as needed */
-    font-family: var(--primary-font), var(--secondary-font);
-}
-
-.create-request-component-student .file-select{
-    left: 1px;
-    margin-top: 7px;
-    position: absolute;
-}
-
-.create-request-component-student .close-button{
-    top: 500px;
-    left: 190px;
-    font-size: large;
-    position: absolute;
-    background-color: var(--primary);
-    color: var(--background);
-    border-radius: 7px;
-    margin-right: 10px;
-    margin-left: auto;
-}
-
-.create-request-component-student .close-button:hover {
-    background-color: var(--accent);
-  }
-
-.create-request-component-student .x-button{
-    top: 0;
-    left: 0;
-    font-size: large;
-    position: absolute;
-    color: #000;
-    background-color: var(--secondary);
-    border-radius: 7px;
-    margin-right: 10px;
-    margin-left: auto;
-}
-
-.create-request-component-student .x-button:hover {
-    background-color: var(--background);
-  }
